add service: Validator, cleanup client services, refactor README.md, add lock file to gitignore
