+++ client/.gitignore
@@ -35,3 +35,5 @@ build/
 node_modules
 
 allure-results
+
+package-lock.json
