+++ client/src/main/java/com/atm/client/Application.java
@@ -1,15 +1,11 @@
-package com.server.server;
+package com.atm.client;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@Slf4j
-public class ServerApplication {
-
+public class Application {
     public static void main(String[] args) {
-        SpringApplication.run(ServerApplication.class, args);
-        log.info("server start");
+        SpringApplication.run(Application.class, args);
     }
 }
