+++ server/src/main/java/com/server/server/Application.java
@@ -1,14 +1,12 @@
-package com.atm.client;
+package com.server.server;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@Slf4j
-public class ClientApplication {
+public class Application {
+
     public static void main(String[] args) {
-        SpringApplication.run(ClientApplication.class, args);
-        log.info("Client start");
+        SpringApplication.run(Application.class, args);
     }
 }
